VMS-7876: Fixed detected objects lookup query after improvements. It was missing objects having track not fully within user-provided filter
VMS-7876: Improving detected objects lookup query performance
VMS-8223: Doubled results of ec2/analyticsLookupDetectedObjects
VMS-8241: metadata_sdk: Added support for building the SDK sample on Nvidia Tegra
VMS-8249: Analytics, actions: added params cameraId, timestampUs to action execution
VMS-8651: Refactored and improved retrieving thumbnails; added .ini option for primary liveCache
VMS-8666: Remove unnecessary information from object tiles in right panel
VMS-8692: If attribute was detected not instantly - show it later (update tile)

<==== Mercurial data ====>
Branch: vms

+++ common_libs/nx_sdk/src/nx/sdk/metadata/common_uncompressed_video_frame.cpp
@@ -60,7 +60,7 @@ int CommonUncompressedVideoFrame::dataSize(int plane) const
     if (!m_externalData.empty())
         return m_externalDataSize.at(plane);
     if (!m_ownedData.empty())
-        return m_ownedData.at(plane).size();
+        return (int) m_ownedData.at(plane).size();
 
     return 0;
 }
